requirements.txt
aiogram==3.18.0
python-dotenv==1.0.1
pydantic==2.10.6
pydantic-settings==2.8.1
certifi>=2023.7.22
asyncpg>=0.30.0
aiohttp>=3.9.0
